requirements.txt
fastapi
uvicorn
pymysql
bcrypt
ython-jose[cryptography]
passlib